feat: Editor move API 구현 및 관련 문서 정리
1. move API를 추가하고 문서/블록 이동 요청을 공통화하여 처리하도록 구현하였습니다.
  - 에디터 오케스트레이터가 EditorMoveResourceType에 따른 라우팅 및 실행을 담당합니다.
  - 문서/블록 이동은 기존 이동 알고리즘을 재사용하도록 정리하였습니다.
2. 관련 문서들을 현재 구조에 맞게 변경하였습니다.

+++ documents-api/src/main/java/com/documents/api/editor/dto/EditorMoveOperationRequest.java
@@ -2,6 +2,8 @@ package com.documents.api.editor.dto;
 
 import java.util.UUID;
 
+import com.documents.service.editor.EditorMoveResourceType;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.NotNull;
@@ -33,7 +35,7 @@ public class EditorMoveOperationRequest {
     private UUID beforeId;
 
     @Schema(description = "클라이언트가 기준으로 삼은 블록 버전", example = "3", nullable = true)
-    private Integer version;
+    private Long version;
 
     @AssertTrue(message = "resourceType이 BLOCK이면 version이 필요합니다.")
     public boolean hasVersionWhenBlockMove() {
